requirements.txt
fastapi==0.122.0
uvicorn==0.38.0
python-multipart==0.0.20
httpx==0.27.0
openai==2.8.1
python-dotenv==1.2.1
